RF: move more common functionality from setup.py to setup_support.py
